+++ common/src/main/java/io/github/mortuusars/scholar/client/gui/screen/BookSigningScreen.java
@@ -5,6 +5,7 @@ import io.github.mortuusars.scholar.Config;
 import io.github.mortuusars.scholar.Scholar;
 import io.github.mortuusars.scholar.book.BookSignature;
 import io.github.mortuusars.scholar.client.gui.Widgets;
+import io.github.mortuusars.scholar.client.gui.screen.edit.SpreadBookEditScreen;
 import io.github.mortuusars.scholar.client.gui.widget.textbox.display.HorizontalAlignment;
 import io.github.mortuusars.scholar.client.gui.widget.textbox.text.FormattedString;
 import io.github.mortuusars.scholar.client.gui.widget.textbox.TextBox;
@@ -29,13 +30,14 @@ import java.util.function.Consumer;
 
 public class BookSigningScreen extends Screen {
     public static final ResourceLocation TEXTURE = Scholar.resource("textures/gui/book_signing.png");
+    public static final ResourceLocation TEXTURE_GOLDEN = Scholar.resource("textures/gui/book_signing_golden.png");
 
     @NotNull
     protected final Minecraft minecraft;
     @NotNull
     protected final Player player;
 
-    protected final Screen parentScreen;
+    protected final SpreadBookEditScreen parentScreen;
     protected final int bookColor;
     protected final Consumer<BookSignature> onSign;
 
@@ -55,7 +57,7 @@ public class BookSigningScreen extends Screen {
     protected String titleText = "";
     protected String authorText = Objects.requireNonNull(Minecraft.getInstance().player).getScoreboardName();
 
-    public BookSigningScreen(Screen parentScreen, int bookColor, Consumer<BookSignature> onSign) {
+    public BookSigningScreen(SpreadBookEditScreen parentScreen, int bookColor, Consumer<BookSignature> onSign) {
         super(Component.empty());
         this.parentScreen = parentScreen;
         this.bookColor = bookColor;
@@ -111,7 +113,7 @@ public class BookSigningScreen extends Screen {
         addRenderableWidget(authorTextBox);
 
         // SIGN
-        signButton = new ImageButton(leftPos + 46, topPos + 108, 22, 22, 149, 0,
+        signButton = new ImageButton(leftPos + 46, topPos + 109, 22, 22, 149, 0,
                 22, TEXTURE, textureWidth, textureHeight,
                 b -> sign(), Component.translatable("book.finalizeButton"));
         MutableComponent component = Component.translatable("book.finalizeButton")
@@ -120,7 +122,7 @@ public class BookSigningScreen extends Screen {
         addRenderableWidget(signButton);
 
         // CANCEL
-        cancelSigningButton = new ImageButton(leftPos + 83, topPos + 108, 22, 22, 171, 0,
+        cancelSigningButton = new ImageButton(leftPos + 82, topPos + 108, 22, 22, 171, 0,
                 22, TEXTURE, textureWidth, textureHeight,
                 b -> cancelSigning(), CommonComponents.GUI_CANCEL);
         cancelSigningButton.setTooltip(Tooltip.create(CommonComponents.GUI_CANCEL));
@@ -169,10 +171,16 @@ public class BookSigningScreen extends Screen {
 
         renderBackground(guiGraphics);
 
+        if (parentScreen.isGolden()) {
+            guiGraphics.blit(TEXTURE_GOLDEN, leftPos, topPos, 0, 0, 0,
+                  imageWidth, imageHeight, textureWidth, textureHeight);
+        } else {
+
         RenderUtil.withColorMultiplied(bookColor, () -> {
             guiGraphics.blit(TEXTURE, leftPos, topPos, 0, 0, 0,
                   imageWidth, imageHeight, textureWidth, textureHeight);
         });
+        }
 
         guiGraphics.blit(TEXTURE, leftPos, topPos + 31, 0, 0, 180,
               imageWidth, 76, textureWidth, textureHeight);